Move fmt utilities to zebra_chain::fmt
